Remove nutjs dependency.  Finish sidecar implementation.

+++ ControlR.Shared/Dtos/SidecarDtos/MouseButtonEventDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace ControlR.Shared.Dtos.SidecarDtos;
 
-public record MouseButtonEventDto(int X, int Y, int Button, bool IsPressed) : SidecarDtoBase(SidecarDtoType.MouseButtonEvent);
+public record MouseButtonEventDto(double X, double Y, int Button, bool IsPressed) : SidecarDtoBase(SidecarDtoType.MouseButtonEvent);
