+++ ControlR.Shared/Dtos/SidecarDtos/MovePointerDto.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json.Serialization;
 
 namespace ControlR.Shared.Dtos.SidecarDtos;
-public record MovePointerDto(int X, int Y, MovePointerType MoveType) : SidecarDtoBase(SidecarDtoType.MovePointer);
+public record MovePointerDto(double X, double Y, MovePointerType MoveType) : SidecarDtoBase(SidecarDtoType.MovePointer);
 
 [JsonConverter(typeof(JsonStringEnumConverter))]
 public enum MovePointerType
